move datatypes to its own module and create consts
